requirements.txt
acme==0.31.0
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
blinker==1.4
boto3==1.10.28
botocore==1.13.28
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
click==6.7
cloud-init==19.2
colorama==0.3.7
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
dataclasses==0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.2.4
django-cors-headers==3.5.0
django-filter==2.2.0
django-storages==1.8
djangorestframework==3.10.3
djongo==1.2.35
dnspython==1.16.0
docutils==0.15.2
fake-useragent==0.1.11
future==0.15.2
gunicorn==19.9.0
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Jinja2==2.10
jmespath==0.9.4
josepy==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lxml==4.4.1
MarkupSafe==1.0
mock==2.0.0
ndg-httpsclient==0.4.4
netifaces==0.10.4
numpy==1.17.4
oauthlib==2.0.6
opencv-python==4.1.2.30
PAM==0.4.2
parsedatetime==2.4
parsel==1.5.2
pbr==3.1.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pycrypto==2.6.1
PyDispatcher==2.0.5
pygobject==3.26.1
PyHamcrest==1.9.0
PyICU==1.9.8
PyJWT==1.5.3
pymongo==3.9.0
pyOpenSSL==19.0.0
pyparsing==2.2.0
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.0
python-debian==0.1.32
python-decouple==3.3
pytz==2019.2
pyxdg==0.25
PyYAML==3.12
queuelib==1.5.0
requests==2.18.4
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
s3transfer==0.2.1
Scrapy==1.7.3
scrapy-fake-useragent==1.1.0
scrapy-splash==0.7.2
scrapyjs==0.2
SecretStorage==2.3.1
service-identity==18.1.0
six==1.12.0
sqlparse==0.2.4
ssh-import-id==5.7
systemd-python==234
Twisted==19.7.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
virtualenv==16.7.5
w3lib==1.21.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.6.0
